requirements.txt
amqp==5.3.1
annotated-types 
anyio 
appdirs==1.4.4
billiard==4.2.1
Brotli 
cairocffi==1.7.1
CairoSVG==2.7.1
celery==5.4.0
certifi 
cffi 
charset-normalizer 
click 
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cssselect2==0.7.0
defusedxml==0.7.1
dnspython==2.7.0
email_validator==2.2.0
fastapi 
fastapi-cli==0.0.7
gevent 
greenlet 
h11 
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huey==2.5.2
idna 
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
kombu==5.4.2
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
orjson==3.10.13
pillow==11.1.0
prompt_toolkit==3.0.48
pycparser 
pydantic 
pydantic-extra-types==2.10.1
pydantic-settings==2.7.1
pydantic_core 
pyee==11.1.1
Pygments==2.19.0
pyppeteer==2.0.0
PySocks 
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
redis==5.2.1
reportlab 
requests 
rich==13.9.4
rich-toolkit==0.12.0
shellingham==1.5.4
six==1.17.0
sniffio 
starlette 
svglib==1.5.1
tinycss2==1.4.0
tqdm==4.67.1
typer==0.15.1
typing_extensions 
tzdata==2024.2
ujson==5.10.0
urllib3 
uvicorn 
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.3
wcwidth==0.2.13
webencodings==0.5.1
websockets==10.4
zipp==3.21.0
zope.event 
zope.interface
playwright
